split password resetter into username and email
