+++ dashboard/app/models/user.rb
@@ -1720,10 +1720,10 @@ class User < ApplicationRecord
 
   def self.send_reset_password_instructions(attributes = {})
     # override of Devise method
-    if RequestStore.store[:current_user]&.admin? & attributes[:username].present?
-      Services::User::PasswordResetter.call(username: attributes[:username])
+    if attributes[:username].present? && RequestStore.store[:current_user]&.admin?
+      Services::User::PasswordResetterByUsername.call(username: attributes[:username])
     else
-      Services::User::PasswordResetter.call(email: attributes[:email])
+      Services::User::PasswordResetterByEmail.call(email: attributes[:email])
     end
   end
