+++ dashboard/test/lib/services/user/password_resetter_by_email_test.rb
@@ -1,37 +1,12 @@
 require 'test_helper'
 
-class Services::User::PasswordResetterTest < ActiveSupport::TestCase
-  let(:reset_via_email) {described_class.new(email: email)}
-  let(:reset_via_username) {described_class.new(username: username)}
+class Services::User::PasswordResetterByEmailTest < ActiveSupport::TestCase
+  let(:described_instance) {described_class.new(email: email)}
 
   let(:email) {Faker::Internet.unique.email}
-  let(:username) {Faker::Internet.unique.username(specifier: 5..19)}
 
-  describe '#call with username' do
-    subject(:reset_password) {reset_via_username.call}
-
-    let(:mail) {ActionMailer::Base.deliveries.first}
-
-    context 'for username with an existing user' do
-      let!(:user) {create(:user, email: email)}
-      before do
-        user.username = username
-        user.save!
-      end
-
-      it 'does not send password reset' do
-        original_user_id = user.id
-        user = reset_password
-
-        _(mail).must_be_nil
-        user.id.must_equal original_user_id
-        user.raw_token.wont_be_nil
-      end
-    end
-  end
-
-  describe '#call with email' do
-    subject(:reset_password) {reset_via_email.call}
+  describe '#call' do
+    subject(:reset_password) {described_instance.call}
 
     let(:mail) {ActionMailer::Base.deliveries.first}
